# server/docs/swagger.yaml
definitions:
  data.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      err_msg:
        type: string
    type: object
  data.ItemVO:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
info:
  contact: {}
paths:
  /comment-ms/v1/{client}/items:
    post:
      consumes:
      - application/json
      description: Create an item record.
      parameters:
      - description: Item information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/data.ItemVO'
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/data.ItemVO'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/data.BaseResponse'
            - properties:
                data:
                  type: string
              type: object
      summary: Item Create
      tags:
      - Item
  /comment-ms/v1/{client}/items/{item_id}:
    get:
      description: Get item information by item ID.
      parameters:
      - description: Client identifier
        enum:
        - customer
        - merchant
        in: path
        name: client
        required: true
        type: string
      - description: Item.ID
        in: path
        name: item_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: Item Query with ID
      tags:
      - Item
swagger: "2.0"